NodeRed-MobileNet-11111111Project/requirement.txt
tensorflow
opencv-python
numpy
paho-mqtt
